Move other block things from State -> VM
