+++ evm/vm/forks/frontier/__init__.py
@@ -3,6 +3,10 @@ from evm.rlp.blocks import BaseBlock  # noqa: F401
 from evm.vm.state import BaseState  # noqa: F401
 
 
+from evm.constants import (
+    BLOCK_REWARD,
+    UNCLE_DEPTH_PENALTY_FACTOR,
+)
 from evm.vm import VM
 from evm.rlp.receipts import (
     Receipt,
@@ -63,3 +67,17 @@ class FrontierVM(VM):
     configure_header = configure_frontier_header
     make_receipt = staticmethod(make_frontier_receipt)
     validate_transaction_against_header = validate_frontier_transaction_against_header
+
+    @staticmethod
+    def get_block_reward():
+        return BLOCK_REWARD
+
+    @staticmethod
+    def get_uncle_reward(block_number, uncle):
+        return BLOCK_REWARD * (
+            UNCLE_DEPTH_PENALTY_FACTOR + uncle.block_number - block_number
+        ) // UNCLE_DEPTH_PENALTY_FACTOR
+
+    @classmethod
+    def get_nephew_reward(cls):
+        return cls.get_block_reward() // 32
